refactor stats logic into class
