+++ src/data/ProgramsD2Repository.ts
@@ -1,5 +1,4 @@
 import _ from "lodash";
-import fs from "fs";
 import { Async } from "domain/entities/Async";
 import { Id } from "domain/entities/Base";
 import { ProgramExport } from "domain/entities/ProgramExport";
@@ -8,31 +7,28 @@ import { D2Api } from "types/d2-api";
 import log from "utils/log";
 import { promiseMap, runMetadata } from "./dhis2-utils";
 import { D2ProgramRules } from "./d2-program-rules/D2ProgramRules";
-import { Stats } from "domain/entities/Stats";
+import { D2Tracker } from "./D2Tracker";
 
 type MetadataRes = { date: string } & { [k: string]: Array<{ id: string }> };
 
-type TrackerDataKey = "events" | "enrollments" | "trackedEntities";
-
-const initialStats = {
-    recordsSkipped: [],
-    errorMessage: "",
-    created: 0,
-    ignored: 0,
-    updated: 0,
-};
-
 export class ProgramsD2Repository implements ProgramsRepository {
-    constructor(private api: D2Api) {}
+    private d2Tracker: D2Tracker;
+
+    constructor(private api: D2Api) {
+        this.d2Tracker = new D2Tracker(this.api);
+    }
 
     async export(options: { ids: Id[]; orgUnitIds: Id[] | undefined }): Async<ProgramExport> {
         const { ids: programIds, orgUnitIds } = options;
         const metadata = await this.getMetadata(programIds);
 
         const getOptions = { programIds, orgUnitIds };
-        const events = await this.getFromTracker<object>("events", getOptions);
-        const enrollments = await this.getFromTracker<D2Enrollment>("enrollments", getOptions);
-        const trackedEntities = await this.getFromTracker<D2TrackedEntity>("trackedEntities", getOptions);
+        const events = await this.d2Tracker.getFromTracker<object>("events", getOptions);
+        const enrollments = await this.d2Tracker.getFromTracker<D2Enrollment>("enrollments", getOptions);
+        const trackedEntities = await this.d2Tracker.getFromTracker<D2TrackedEntity>(
+            "trackedEntities",
+            getOptions
+        );
 
         /* Remove redundant enrollments info from TEIs */
         const trackedEntitiesWithoutEnrollments = trackedEntities.map(trackedEntity => ({
@@ -86,116 +82,16 @@ export class ProgramsD2Repository implements ProgramsRepository {
         }));
 
         log.info(`Import data`);
-        await this.postTracker("trackedEntities", trackedEntities);
-        await this.postTracker("enrollments", enrollmentsWithAttributes);
-        await this.postTracker("events", events);
+        await this.d2Tracker.postTracker("trackedEntities", trackedEntities);
+        await this.d2Tracker.postTracker("enrollments", enrollmentsWithAttributes);
+        await this.d2Tracker.postTracker("events", events);
     }
 
     async runRules(options: RunRulesOptions): Async<void> {
         const d2ProgramRules = new D2ProgramRules(this.api);
         return d2ProgramRules.run(options);
     }
-
-    /* Private */
-
-    async postTracker(key: TrackerDataKey, objects: object[]): Async<TrackerResponse[]> {
-        const total = objects.length;
-        log.info(`Import data: ${key} - Total: ${total}`);
-        let page = 1;
-        const chunkSize = 100;
-        const result = [];
-
-        for (const objectsGroup of _.chunk(objects, chunkSize)) {
-            const initial = (page - 1) * chunkSize;
-            log.debug(`Import data (${objectsGroup.length}): ${key} (offset: ${initial})`);
-            const trackerResponse = await this.postTrackerData(
-                { [key]: objectsGroup },
-                { payloadId: `${key}-${initial}` }
-            );
-            result.push(trackerResponse);
-            page++;
-        }
-        return result;
-    }
-
-    async postTrackerData(data: object, options: { payloadId: string }): Async<TrackerResponse> {
-        const response: TrackerResponse = await this.api
-            .post<TrackerResponse>("/tracker", { async: false }, data)
-            .getData()
-            .catch(err => {
-                if (err?.response?.data) {
-                    return err.response.data as TrackerResponse;
-                } else {
-                    return {
-                        status: "ERROR",
-                        typeReports: [],
-                        stats: initialStats,
-                    };
-                }
-            });
-        log.debug(`Status: ${response.status}`);
-
-        if (response.status !== "OK") {
-            const errorJsonPath = `programs-import-error-${options.payloadId}.json`;
-            log.error(`Save import error: ${errorJsonPath}`);
-            fs.writeFileSync(errorJsonPath, JSON.stringify({ response, data }, null, 4));
-            return response;
-        } else {
-            return response;
-        }
-    }
-
-    async getFromTracker<T>(
-        apiPath: string,
-        options: {
-            programIds: string[];
-            orgUnitIds: string[] | undefined;
-            fields?: string;
-            trackedEntity?: string | undefined;
-        }
-    ): Promise<T[]> {
-        const output = [];
-        const { programIds, orgUnitIds, fields = "*", trackedEntity } = options;
-
-        for (const programId of programIds) {
-            let page = 1;
-            let dataRemaining = true;
-
-            while (dataRemaining) {
-                const pageSize = 1000;
-                log.debug(`GET ${apiPath} (pageSize=${pageSize}, page=${page})`);
-
-                const { instances } = await this.api
-                    .get<{ instances: T[] }>(`/tracker/${apiPath}`, {
-                        page,
-                        pageSize: pageSize,
-                        ouMode: orgUnitIds ? "SELECTED" : "ALL",
-                        orgUnit: orgUnitIds?.join(";"),
-                        fields: fields,
-                        program: programId,
-                        trackedEntity,
-                    })
-                    .getData();
-
-                if (instances.length === 0) {
-                    dataRemaining = false;
-                } else {
-                    output.push(...instances);
-                    page++;
-                }
-            }
-        }
-        log.info(`GET ${apiPath} -> Total: ${output.length}`);
-
-        return output;
-    }
 }
-
-type TrackerResponse = {
-    status: string;
-    typeReports: object[];
-    stats: Stats;
-};
 
 interface D2ProgramExport {
     metadata: object;
