+++ src/data/TrackedEntityD2Repository.ts
@@ -9,31 +9,21 @@ import {
     TrackedEntityRepository,
 } from "domain/repositories/TrackedEntityRepository";
 import { TrackedEntity } from "domain/entities/TrackedEntity";
-import { D2TrackedEntity, ProgramsD2Repository } from "./ProgramsD2Repository";
-
-const initialStats = {
-    recordsSkipped: [],
-    errorMessage: "",
-    created: 0,
-    ignored: 0,
-    updated: 0,
-};
+import { D2TrackedEntity } from "./ProgramsD2Repository";
+import { D2Tracker } from "./D2Tracker";
 
 export class TrackedEntityD2Repository implements TrackedEntityRepository {
-    private programsD2Repository: ProgramsD2Repository;
+    private d2Tracker: D2Tracker;
 
     constructor(private api: D2Api) {
-        this.programsD2Repository = new ProgramsD2Repository(this.api);
+        this.d2Tracker = new D2Tracker(this.api);
     }
 
     async getAll(params: TrackedEntityFilterParams): Async<TrackedEntity[]> {
-        const trackedEntities = await this.programsD2Repository.getFromTracker<D2TrackedEntity>(
-            "trackedEntities",
-            {
-                orgUnitIds: undefined,
-                programIds: [params.programId],
-            }
-        );
+        const trackedEntities = await this.d2Tracker.getFromTracker<D2TrackedEntity>("trackedEntities", {
+            orgUnitIds: undefined,
+            programIds: [params.programId],
+        });
 
         return trackedEntities.map(tei => {
             return {
@@ -52,28 +42,24 @@ export class TrackedEntityD2Repository implements TrackedEntityRepository {
         });
     }
 
-    async saveAttributes(teis: TrackedEntity[]): Async<Stats> {
-        if (teis.length === 0)
-            return { created: 0, ignored: 0, updated: 0, errorMessage: "", recordsSkipped: [] };
-        const teisToFetch = teis.map(program => program.id);
-        const programsIds = _(teis)
+    async save(trackedEntities: TrackedEntity[]): Async<Stats> {
+        if (trackedEntities.length === 0) return Stats.empty();
+        const teisToFetch = trackedEntities.map(program => program.id);
+        const programsIds = _(trackedEntities)
             .map(program => program.programId)
             .uniq()
             .value();
 
-        const programsByKey = _(teis)
+        const programsByKey = _(trackedEntities)
             .keyBy(tei => tei.id)
             .value();
 
         const stats = await getInChunks<Stats>(teisToFetch, async teiIds => {
-            const trackedEntities = await this.programsD2Repository.getFromTracker<D2TrackedEntity>(
-                "trackedEntities",
-                {
-                    orgUnitIds: undefined,
-                    programIds: programsIds,
-                    trackedEntity: teiIds.join(";"),
-                }
-            );
+            const trackedEntities = await this.d2Tracker.getFromTracker<D2TrackedEntity>("trackedEntities", {
+                orgUnitIds: undefined,
+                programIds: programsIds,
+                trackedEntity: teiIds.join(";"),
+            });
 
             const teisToSave = trackedEntities.map(tei => {
                 const currentProgram = programsByKey[tei.trackedEntity];
@@ -90,30 +76,22 @@ export class TrackedEntityD2Repository implements TrackedEntityRepository {
                 };
             });
 
-            const response = await this.programsD2Repository.postTracker("trackedEntities", teisToSave);
+            const response = await this.d2Tracker.postTracker("trackedEntities", teisToSave);
 
             return _(response)
                 .map(item => {
                     const isError = item.status === "ERROR";
-                    return {
+                    return new Stats({
                         created: item.stats.created,
                         updated: item.stats.updated,
                         ignored: item.stats.ignored,
                         errorMessage: isError ? item.status : "",
                         recordsSkipped: isError ? teisToSave.map(tei => tei.trackedEntity) : [],
-                    };
+                    });
                 })
                 .value();
         });
 
-        return stats.reduce((acum, stat) => {
-            return {
-                recordsSkipped: [...acum.recordsSkipped, ...stat.recordsSkipped],
-                errorMessage: `${acum.errorMessage}${stat.errorMessage}`,
-                created: acum.created + stat.created,
-                ignored: acum.ignored + stat.ignored,
-                updated: acum.updated + stat.updated,
-            };
-        }, initialStats);
+        return Stats.combine(stats);
     }
 }
